requirements.txt
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==4.6.0
django-filter==24.3
django-model-utils==5.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ>=0.11.2
django-extensions>=3.2.3
django-storages>=1.14.2
django-redis>=5.4.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
celery>=5.3.4
Pillow>=10.1.0
pytz>=2023.3.post1
redis>=5.0.1
requests>=2.31.0
whitenoise>=6.6.0
pgvector>=0.2.3
python-magic>=0.4.27
pdf2image>=1.16.3
pypdf==5.4.0
langchain>=0.0.350
langchain-community>=0.0.13
langchain-nvidia-ai-endpoints==0.3.9
faiss-cpu>=1.7.4
openai>=1.3.7
langchain-openai>=0.0.8
tiktoken
tqdm>=4.66.1
dj-rest-auth==6.0.0
qdrant-client>=1.7.0